Fixed signed/unsigned problems
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@1058 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ external/lockfree/boost/lockfree/detail/gcc_primitives.hpp
@@ -286,10 +286,10 @@ namespace boost { namespace lockfree
 
     inline bool interlocked_bit_test_and_set(long* x, long bit)
     {
-        long const value = 1 << bit;
+        unsigned long const value = 1u << bit;
         long old = *x;
         do {
-            long const current = interlocked_compare_exchange(x, old, old | value);
+            long const current = interlocked_compare_exchange(x, old, long(old | value));
             if (current == old)
             {
                 break;
@@ -301,10 +301,10 @@ namespace boost { namespace lockfree
 
     inline bool interlocked_bit_test_and_reset(long* x, long bit)
     {
-        long const value = 1 << bit;
+        unsigned long const value = 1u << bit;
         long old = *x;
         do {
-            long const current = interlocked_compare_exchange(x, old, old & ~value);
+            long const current = interlocked_compare_exchange(x, old, long(old & ~value));
             if (current == old)
             {
                 break;
